code refactor and new models for #15
